Add active_model_serializers gem
Not super happy with the rocket_pants/active_model_serializer combo,
but it works for now. Mostly, I want to be able to follow the
jsonapi.org standard because it makes enough sense to me and means I'm
at least following some sort of standard. I couldn't get the
active_model_serializers 0.10 rc, which apparently supports jsonapi, to
work with sequel.